Refactor `msuac` class to use `registry::value`
Without this commit, the `msuac` class was declaring directly a
different type for the registry key and the associated value. This is
unnecessary since there is the `registry::value` defined type that will
control the value and optionally control the key existence if it is not
yet declared.

This commit refactors the `msuac` class to use the `registry::value`
defined type in favor of separate declarations. It also uses different
names for the variables the get set to the value data in order to make
the usage more clear.