Add failure on non-windows systems
Without this commit, this class could become part of a catalog for a
system that is not a windows machine. This catalog could then be
partially applied but would fail on the resources within this class.

This commit adds a check that will fail catalog compilation in the event
this class is applied to a non-windows machine.